kernel: sched: hold scheduler lock when signaling IPIs
signal_pending_ipi() must run under _sched_spinlock so flag_ipi()
cannot race the pending_ipi drain. z_reschedule_irqlock() had only
the IRQ lock. reschedule() had the caller's lock, which is often an
object lock rather than _sched_spinlock.

Take the scheduler lock on both no-swap paths. Use
LOCK_SCHED_SPINLOCK so the section is empty when
CONFIG_MP_MAX_NUM_CPUS is 1. If reschedule() holds a different lock,
release it first (IRQs stay masked), matching do_swap().

Tested on qemu_cortex_m3 (cmake/ninja): schedule_api 21 pass / 8
skip, thread_apis 41 pass / 2 skip. Skips are ZTEST_USER and SMP
cpu_mask; this board has USERSPACE=n.

+++ drivers/disk/flashdisk.c
@@ -126,8 +126,7 @@ static int flashdisk_init_runtime(struct flashdisk_data *ctx,
 		while (offset < ctx->offset + ctx->size) {
 			rc = flash_get_page_info_by_offs(ctx->info.dev, offset, &page);
 			if (rc < 0) {
-				LOG_ERR("Error %d getting page info at offset %lx", rc,
-					(long)offset);
+				LOG_ERR("Error %d getting page info at offset %lx", rc, offset);
 				return rc;
 			}
 			if (page.size != ctx->page_size) {
